+++ backend/src/RiskAnalyzer.Api/Models/ImageAnalysis.cs
@@ -9,10 +9,10 @@ public class ImageAnalysis
     public string FilePath { get; set; } = null!;
     public long FileSizeBytes { get; set; }
     public string ContentType { get; set; } = null!;
-    public int RiskScore { get; set; }
+    public decimal RiskScore { get; set; }
     public string Classification { get; set; } = null!;
     public string AnalysisDetails { get; set; } = null!;
-    public int ConfidenceScore { get; set; }
+    public decimal ConfidenceScore { get; set; }
     public DateTime AnalyzedAt { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public string? Categories { get; set; }
